+++ src/initial/initial_field_coefficients_wavebundle_normal_incidence.m
@@ -80,8 +80,7 @@ gaussfacSincos = gaussfac.*cb.*sb;
 [pilm,taulm] = spherical_functions_trigon(cb,sb,lmax);  % Nk x 1
 
 % cylindrical coordinates for relative particle positions
-relativeParticlePositions = bsxfun(@plus,simulation.input.particles.positionArray, ...
-                                        -simulation.input.initialField.focalPoint);
+relativeParticlePositions = simulation.input.particles.positionArray - simulation.input.initialField.focalPoint;
 rhoGi = sqrt(relativeParticlePositions(:,1).^2+relativeParticlePositions(:,2).^2); % NS x 1
 phiGi = atan2(relativeParticlePositions(:,2),relativeParticlePositions(:,1)); % NS x 1
 zGi = relativeParticlePositions(:,3); % NS x 1
@@ -94,8 +93,8 @@ for m=-lmax:lmax
     eikzJmm1 = eikz .* besselj(abs(m-1),rhoGi*k*sb.'); % NS x Nk
     eimp1phi = exp(-1i*(m+1)*phiGi); % NS x 1
     eimm1phi = exp(-1i*(m-1)*phiGi); % NS x 1
-    plTerm = exp(1i*alphaG)*1i^abs(m+1)*bsxfun(@times,eimp1phi,eikzJmp1);
-    mnTerm = exp(-1i*alphaG)*1i^abs(m-1)*bsxfun(@times,eimm1phi,eikzJmm1);
+    plTerm = exp(1i*alphaG)*1i^abs(m+1)*(eimp1phi.*eikzJmp1);
+    mnTerm = exp(-1i*alphaG)*1i^abs(m-1)*(eimm1phi.*eikzJmm1);
     eikzI1 = pi*( mnTerm + plTerm );
     eikzI2 = pi*1i*(-mnTerm + plTerm);
     for tau=1:2