+++ src/postprocessing/scattered_field_plane_wave_pattern.m
@@ -52,7 +52,13 @@ alphaArray = simulation.numerics.azimuthalAnglesArray;
 alphaDim = length(alphaArray);
 betaDim = length(betaArray);
 
-pwp(1:2) = {celes_planeWavePattern(betaArray, alphaArray, simulation.input.k_medium)};
+for pol = 2:-1:1
+    pwp{pol} = celes_planeWavePattern('polarAngles', betaArray, ...
+                                      'azimuthalAngles', alphaArray, ...
+                                      'k', simulation.input.k_medium, ...
+                                      'expansionCoefficients', simulation.numerics.deviceArray(zeros(alphaDim,betaDim,'single')) ... % dima x dimb
+                                      );
+end
 
 kxGrid = pwp{1}.kxGrid;
 kyGrid = pwp{1}.kyGrid;
@@ -60,7 +66,9 @@ kzGrid = pwp{1}.kzGrid;
 
 [pilm,taulm] = spherical_functions_trigon(cb,sb,lmax); % 1 x dimb
 
-B(1:2) = {simulation.numerics.deviceArray(zeros(nmax,betaDim,'single'))}; % dimn x dimb
+for pol = 2:-1:1
+    B{pol} = simulation.numerics.deviceArray(zeros(nmax,betaDim,'single')); % dimn x dimb
+end
 
 for tau = 1:2
     for l = 1:lmax
@@ -83,10 +91,6 @@ for tau = 1:2
     end
 end
 
-for pol = 1:2
-    pwp{pol}.expansionCoefficients = simulation.numerics.deviceArray(zeros(alphaDim,betaDim,'single')); % dima x dimb
-end
-
 for jS = 1:simulation.input.particles.number
     fprintf(1,repmat('\b',1,length(msg)));
     msg = sprintf(' sphere %i of %i',jS,simulation.input.particles.number);
@@ -95,7 +99,7 @@ for jS = 1:simulation.input.particles.number
     emnirk = exp(-1i*(simulation.input.particles.positionArray(jS,1)*kxGrid+ ...
                       simulation.input.particles.positionArray(jS,2)*kyGrid+ ...
                       simulation.input.particles.positionArray(jS,3)*kzGrid)); % dima x dimb
-    beima = bsxfun(@times,eima,simulation.tables.scatteredFieldCoefficients(jS,:)); % dima x dimn
+    beima = eima.*simulation.tables.scatteredFieldCoefficients(jS,:); % dima x dimn
     for pol = 1:2
         pwp{pol}.expansionCoefficients = pwp{pol}.expansionCoefficients+ ...
                                         (beima * B{pol}).* emnirk / (2*pi);