+++ CHANGELOG.md
@@ -13,15 +13,24 @@
 
 ## [Testing]
 ### Changed
-- subclass celes classes from `matlab.System`
+-
+
+## [Unreleased]
+### Changed
+- subclass CELES classes from `matlab.System`
 - implement `setProperties` methods for name-value style initialization
 - implement `validatePropertiesImpl` methods for validation
 - limit use of `Dependent` properties to avoid redundant calculations
 - implement `setupImpl` where appropriate for one-time calculations to improve performance
 - compute and set maximal particle distance inside particles class
+- compute and set h3_table inside simulation class
+- compute and set Plm_coefficient table inside numerics class
+- initialize simulation instance with a default output instance
+- prefer implicit expansion over `bsxfun` (requires MATLAB >= R2016b)
+- fancier, faster plotting functions
 
-## [Unreleased]
--
+### Fixed
+- remove hardcoded lookup resolution value in GUI example parameters
 
 ## [2.1] - 2017-10-25
 ### Changed